Ch. 18: Show joke in a notification.
The joke instead of being shown in the log, is now shown in a notification.
The version of NotificationCompat.Builder(Context) is now deprecated, and
instead must use NotificationCompat.Builder(Context, String), where the
second argument is a channel ID. The link to this channel setup code is
in the code.